backend/requirements.txt
fastapi
uvicorn
google-generativeai
pandas
faker
python-multipart
python-dotenv
